refactored many classes: Exchanged old TPTP type by new TPTPAST. First runs seem to work

+++ src/main/scala/leo/datastructures/impl/orderings/ClauseProxyOrderings.scala
@@ -103,7 +103,7 @@ object CPO_OldestFirst extends ClauseProxyOrdering[Double] {
     aAge.compareTo(bAge)
   }
 
-  final def weightOf(cl: ClauseProxy): Double = cl.id
+  final def weightOf(cl: ClauseProxy): Double = cl.id.toDouble
 }
 
 /** Ordering in which a [[leo.datastructures.ClauseProxy]] is smaller, if its ratio of negative literals is greater.
